Added a FlipTileSkin
The FlipTileSkin looks and behaves like a so called SplitFlap which was used a lot in train stations for the departure and arrival boards.
This Skin will make use of new properties like
- flipTimeInMS (flip time in milliseconds between 0 - 2000 ms)
- flipText (the text that should be shown on the tile)
- characterList (a list of strings which represents the available characters)
To make it more convenient to use I've already added some character sets that might come in handy e.g. 0-59, 0-12, 0-24, A-Z etc.
These sets can be found in the Helper class.